+++ src/background/index.js
@@ -1,0 +1,3 @@
+import { registerTabWatchers } from './modules/tabWatchers.js';
+
+registerTabWatchers();